Get all fedsd utility all in one file, as suggested by code review
